Rewrite useAllTimeBlock to function
